Buoi2/pyproject.toml
[project]
name = "buoi2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.74.0",
    "openai-agents>=0.0.11",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
]
